Data.SOP.Deep

Documentation

type family StepDeep (x :: a) :: [[a]] where ...

Expand a single index to its deep representation

Equations

StepDeep x = StepDeep1 (DCode x)

singIStepDeep :: forall a f (x :: a). HasDeep a f => SingI x :- SingI (StepDeep x)

StepDeep preserves SingI

singStepDeep :: forall a f (x :: a). HasDeep a f => Sing x -> Sing (StepDeep x)

StepDeep preserves Sing

type family StepDeep1 (code :: [[a]]) :: [[a]] where ...

Singletons of singletons are shallow, everything else is deep

Equations

StepDeep1 ((x ': '[]) ': '[]) = (x ': '[]) ': '[]

StepDeep1 code = Deep code

singStepDeep1 :: forall a f (code :: [[a]]). HasDeep a f => Sing code -> Sing (StepDeep1 code)

StepDeep1 preserves Sing

type family Deep (xs :: [[a]]) :: [[a]] where ...

We expand deeply by MapStepDeep1ing and joining the results

Equations

Deep xs = Map ("Concat" :# ()) (Concat (Map ("Distribute" :# ()) (Map ("MapStepDeep" :# ()) xs)))

singDeep :: forall a f (xs :: [[a]]). HasDeep a f => Sing xs -> Sing (Deep xs)

Deep preserves Sing

class HasDict1 a => HasDeep a (f :: a -> Type) | a -> f where

An indexed form of Generic that allows deep conversion to/from SOP

-- | Unsafe un-application
type family UnF (f :: a -> Type) (y :: Type) :: a where
  UnF f (f x) = x
  UnF f x = TypeError ('Text "UnF expected (" ':<>: 'ShowType f ':<>: 'Text " _) but got " ':<>: 'ShowType x)

MapUnF
MapUnF1

-- | If your f is consistently recognized by UnF,
-- GDCode should be equivalent to DCode
type family GDCode (f :: a -> Type) (x :: a) where
  GDCode f x = MapUnF1 f (Code (f x))

Minimal complete definition

Nothing

Associated Types

type DCode (x :: a) :: [[a]]

A Dependent/Deep Code

Methods

singIDCode :: forall (x :: a). SingI x :- SingI (DCode x)

DCode must preserve SingI

singDCode :: forall (x :: a). Sing x -> Sing (DCode x)

DCode must preserve Sing

toD :: forall (x :: a). Sing x -> SOP f (DCode x) -> f x

to for DCode

toD :: forall (x :: a). (Generic (f x), AllZipN (Prod SOP) (LiftedCoercible f I) (DCode x) (Code (f x))) => Sing x -> SOP f (DCode x) -> f x

to for DCode

fromD :: forall (x :: a). Sing x -> f x -> SOP f (DCode x)

from for DCode

fromD :: forall (x :: a). (Generic (f x), AllZipN (Prod SOP) (LiftedCoercible I f) (Code (f x)) (DCode x)) => Sing x -> f x -> SOP f (DCode x)

from for DCode

fromStepDeep :: forall a f (x :: a). HasDeep a f => Sing x -> f x -> SOP f (StepDeep x)

Convert an embedded value to a SOP of StepDeep by conditionally applying fromD or fromDeep depending on the result of singDCode

fromMapStepDeep1 :: forall a f (x :: a). HasDeep a f => Sing x -> SOP f (DCode x) -> SOP (SOP f) (Map ("MapStepDeep" :# ()) (DCode x))

Map fromStepDeep over a SOP of DCode for some value

fromDeep :: forall a f (x :: a). HasDeep a f => Sing x -> f x -> SOP f (Deep (DCode x))

Convert an embedded value to a Deep SOP of its DCode

toStepDeep :: forall a f (x :: a). HasDeep a f => Sing x -> SOP f (StepDeep x) -> f x

Convert a SOP of StepDeep to an embedded value by conditionally applying fromD or fromDeep depending on the result of singDCode

toMapStepDeep1 :: forall a f (x :: a). HasDeep a f => Sing x -> SOP (SOP f) (Map ("MapStepDeep" :# ()) (DCode x)) -> SOP f (DCode x)

Map toStepDeep to a SOP of DCode for some value

toDeep :: forall a f (x :: a). HasDeep a f => Sing x -> SOP f (Deep (DCode x)) -> f x

Convert a Deep SOP of a value's DCode to an embedding of it

class (HasDeep a f, DCode x ~ '['[x]]) => IsDeep f (x :: a)

Iff (DCode x) is a singleton of a singleton, Deep does not recurse on x

Orphan instances

HasDeep a f => HasMap ("StepDeep" :# ()) a

Instance details

Associated Types

type Mapper ("StepDeep" :# ()) x :: MapOut name a

Methods

singIMapper :: SingI x :- SingI (Mapper ("StepDeep" :# ()) x)

singMapper :: Sing x -> Sing (Mapper ("StepDeep" :# ()) x)

HasDeep a f => HasMap ("MapStepDeep" :# ()) [a]

Instance details

Associated Types

type Mapper ("MapStepDeep" :# ()) x :: MapOut name a

Methods

singIMapper :: SingI x :- SingI (Mapper ("MapStepDeep" :# ()) x)

singMapper :: Sing x -> Sing (Mapper ("MapStepDeep" :# ()) x)