Data.SOP.Join

Documentation

mapAssocPOP :: forall a (f :: a -> Type) (xs :: [[[a]]]). HasDict1 a => Sing xs -> NS (POP f) xs -> NS (NP f) (Map ("Concat" :# ()) xs)

Map assocPOP

unMapAssocPOP :: forall a (f :: a -> Type) (xs :: [[[a]]]). HasDict1 a => Sing xs -> NS (NP f) (Map ("Concat" :# ()) xs) -> NS (POP f) xs

Map unAssocPOP

innerToPOS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). Sing xs -> SOP (SOP f) xs -> NS (POS (NP f)) xs

A nested sum of products is isomorphic to a sum of a product-of-sums of a product

innerFromPOS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). Sing xs -> NS (POS (NP f)) xs -> SOP (SOP f) xs

A sum of a product-of-sums of a product is isomorphic to a nested sum of products

innerDistributePOS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (POS (NP f)) xs -> NS (SOP (NP f)) (Map ("Distribute" :# ()) xs)

Map distributePOS

innerUnDistributePOS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (SOP (NP f)) (Map ("Distribute" :# ()) xs) -> NS (POS (NP f)) xs

Map unDistributePOS

innerToNSNP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (SOP (NP f)) (Map ("Distribute" :# ()) xs) -> NS (NS (NP (NP f))) (Map ("Distribute" :# ()) xs)

Expand SOP into a sum of products

innerFromNSNP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (NS (NP (NP f))) (Map ("Distribute" :# ()) xs) -> NS (SOP (NP f)) (Map ("Distribute" :# ()) xs)

Retract a sum of products into SOP

concatNS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (NS (NP (NP f))) (Map ("Distribute" :# ()) xs) -> NS (NP (NP f)) (Concat (Map ("Distribute" :# ()) xs))

Concat the outer NS through association (assocSOS)

unConcatNS :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (NP (NP f)) (Concat (Map ("Distribute" :# ()) xs)) -> NS (NS (NP (NP f))) (Map ("Distribute" :# ()) xs)

Un-concat the outer NS through un-association (unAssocSOS)

concatNP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (NP (NP f)) (Concat (Map ("Distribute" :# ()) xs)) -> NS (NP f) (Map ("Concat" :# ()) (Concat (Map ("Distribute" :# ()) xs)))

Concat the inner NP's through association (mapAssocPOP)

unConcatNP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> NS (NP f) (Map ("Concat" :# ()) (Concat (Map ("Distribute" :# ()) xs))) -> NS (NP (NP f)) (Concat (Map ("Distribute" :# ()) xs))

Un-concat the inner NP's through association (unMapAssocPOP)

joinSOP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> SOP (SOP f) xs -> SOP f (Map ("Concat" :# ()) (Concat (Map ("Distribute" :# ()) xs)))

Join nested SOP's

 \sxs ->
 SOP .
 concatNP sxs .
 concatNS sxs .
 innerToNSNP sxs .
 innerDistributePOS sxs .
 innerToPOS sxs

unJoinSOP :: forall a (f :: a -> Type) (xs :: [[[[a]]]]). HasDict1 a => Sing xs -> SOP f (Map ("Concat" :# ()) (Concat (Map ("Distribute" :# ()) xs))) -> SOP (SOP f) xs

Un-join nested SOP's

 \sxs ->
 innerFromPOS sxs .
 innerUnDistributePOS sxs .
 innerFromNSNP sxs .
 unConcatNS sxs .
 unConcatNP sxs .
 unSOP