Data.SOP.Map

Documentation

type (:#) (name :: Symbol) (a :: k) = NamedF Identity (Sing a) name

class c a b => Flip c b a

singAllTop :: forall a (xs :: [a]). Sing xs -> Dict (All Top xs)

singAllTop2 :: forall a (xs :: [[a]]). Sing xs -> Dict (All (All Top) xs)

type family MapOut (name :: Type) (x :: a) :: Type

The output of Mapper

class (HasDict1 a, HasDict1 (MapOut name a)) => HasMap (name :: Type) (a :: Type) where

A type family that can be mapped

Minimal complete definition

singMapper

Associated Types

type Mapper (name :: Type) (x :: a) :: MapOut name a

A type family that may be mapped

Methods

singIMapper :: forall (x :: a). SingI x :- SingI (Mapper name x)

Mapper preserves SingI

singMapper :: forall (x :: a). Sing x -> Sing (Mapper name x)

Mapper preserves Sing

type family Map (name :: Type) (xs :: [a]) :: [MapOut name a] where ...

Map the Mapper of the given name over the list of inputs

Equations

Map _ '[] = '[]

Map name (x ': xs) = Mapper name x ': Map name xs

singEmptyList :: Dict (SingI '[])

A hack, was getting an overlapping instance error in singIMap

singIMap :: forall (name :: Type) a (xs :: [a]). HasMap name a => SingI xs :- SingI (Map name xs)

Proof that Map preserves SingI (when Mapper does)

singMap :: forall (name :: Type) a (xs :: [a]). HasMap name a => Sing xs -> Sing (Map name xs)

Proof that Map preserves Sing (when Mapper does)

sameShapeMap :: forall name a (xs :: [a]). HasMap name a => Sing xs -> Dict (SameShapeAs xs (Map name xs), SameShapeAs (Map name xs) xs)

Map preserves shape

class (SingI x, Mapper name x ~ y) => IsMapper (name :: Type) (x :: a) (y :: MapOut name a)

Relational form of Mapper for AllZipF

zipMapper :: forall name a (xs :: [a]). HasMap name a => Sing xs -> Dict (All Top xs, All Top (Map name xs), SameShapeAs xs (Map name xs), SameShapeAs (Map name xs) xs, AllZipF (IsMapper name) xs (Map name xs), AllZipF (Flip (IsMapper name)) (Map name xs) xs)

Proof that zipping xs and Map name xs with (IsMapper name) holds

ixmap :: forall name a (f :: a -> Type) (g :: MapOut name a -> Type) h1 h2 (xs :: [a]). (HasMap name a, HTrans h1 h2, Prod h1 ~ NP) => (forall (x :: a). Sing x -> f x -> g (Mapper name x)) -> Sing xs -> h1 f xs -> h2 g (Map name xs)

Map the indices of some HTrans pair

ixunmap :: forall name a (f :: a -> Type) (g :: MapOut name a -> Type) h1 h2 (xs :: [a]). (HasMap name a, HTrans h1 h2, Prod h1 ~ NP) => (forall (x :: a). Sing x -> g (Mapper name x) -> f x) -> Sing xs -> h1 g (Map name xs) -> h2 f xs

Unmap the indices of some HTrans pair