Data.Singletons.Map

Documentation

newtype ListMap k a

A map from k to a implement using a list

Constructors

MkListMap

Fields

unListMap :: [(k, a)]

emptyListMap :: ListMap k a

An empty ListMap

lookupModifyListMap :: Ord k => a -> (a -> (b, a)) -> k -> ListMap k a -> (b, ListMap k a)

Lookup and modify a value, given a default value if it's not in the map.