Data.Singletons.TH.Cases

Description

A simple modification of sCases to allow generating the body with case information: (Name -> [Name] -> q Exp) and some utilities.

 Copyright (c) 2012, Richard Eisenberg
 All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:

 1. Redistributions of source code must retain the above copyright notice, this
 list of conditions and the following disclaimer.

 2. Redistributions in binary form must reproduce the above copyright notice,
 this list of conditions and the following disclaimer in the documentation
 and/or other materials provided with the distribution.

 3. Neither the name of the author nor the names of its contributors may be
 used to endorse or promote products derived from this software without
 specific prior written permission.

 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
 DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
 FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
 DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
 SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Documentation

sCasesFold :: ExpQ -> ExpQ -> Name -> ExpQ

Example:

 foo = $(sCasesFold f ''Either)
 foo = x ->
   case x of
     Left y -> $(qFunc) y $(qBase)
     Right y -> $(qFunc) y $(qBase)

sCases'

Arguments

:: DsMonad q

=> Name
The head of the type the scrutinee's type is based on. (Like ''Maybe or ''Bool.)

-> q Exp
The scrutinee, in a Template Haskell quote

-> (Name -> [Name] -> q Exp)

-> q Exp

The function sCases generates a case expression where each right-hand side is identical. This may be useful if the type-checker requires knowledge of which constructor is used to satisfy equality or type-class constraints, but where each constructor is treated the same. For sCases, unlike cases, the scrutinee is a singleton. But make sure to pass in the name of the original datatype, preferring ''Maybe over ''SMaybe.

buildCases :: forall m. DsMonad m => [(Name, Int)] -> m Exp -> (Name -> [Name] -> m Exp) -> m DExp

Utility for sCases'

tysOfConFields :: DConFields -> [DType]

extractNameArgs :: DCon -> (Name, Int)

Extract the name and number of arguments to a constructor

extractNameTypes :: DCon -> (Name, [DType])

Extract the name and types of constructor arguments

prefixConName :: String -> String -> Name -> Name

singPkg :: String

Singletons package

mk_name_d :: String -> String -> Name

Make singletons name

sconsName :: Name

SCons

snilName :: Name

SNil

mkTupleDataName :: Int -> Name

Tuple singleton Name's

singDataConName :: Name -> Name

Data constructor singleton Name's