Data.Singletons.WrappedSing

Description

Compat module until singletons >= 2.6 can be used with morley

Documentation

newtype WrappedSing a where

A newtype around Sing.

Since Sing is a type family, it cannot be used directly in type class instances. As one example, one cannot write a catch-all instance SDecide k => TestEquality (Sing k). On the other hand, WrappedSing is a perfectly ordinary data type, which means that it is quite possible to define an instance SDecide k => TestEquality (WrappedSing k). type WrappedSing :: k -> Type

Constructors

WrapSing

Fields

:: forall (a :: k). { unwrapSing :: Sing a

} -> WrappedSing a

newtype SWrappedSing ws where

The singleton for WrappedSings. Informally, this is the singleton type for other singletons. type SWrappedSing :: forall k (a :: k). WrappedSing a -> Type

Constructors

SWrapSing

Fields

:: forall (a :: k) (ws :: WrappedSing a). { sUnwrapSing :: Sing a

} -> SWrappedSing ws

type family UnwrapSing (ws :: WrappedSing a) :: Sing a where ...

Equations

UnwrapSing (WrapSing s) = s