Lorentz.Contracts.Util

Documentation

betweenChars :: Char -> Char -> ReadP a -> ReadP a

Parse something between the two given Lexeme's

inParensP :: ReadP a -> ReadP a

Parse something in parentheses

inQuotesP :: ReadP a -> ReadP a

Parse something in double-quotes: "[something]"

maybeLiftP :: (ReadP a -> ReadP a) -> ReadP a -> ReadP a

Attempt to parse with given modifier, otherwise parse without

maybeInParensP :: ReadP a -> ReadP a

Attempt to parse inParensP, else parse without

maybeInQuotesP :: ReadP a -> ReadP a

Attempt to parse inQuotesP, else parse without

readAddressP :: ReadP Address

Read an Address, inside or outside of ""'s

Orphan instances

Ord PublicKey

Instance details

Methods

compare :: PublicKey -> PublicKey -> Ordering

(<) :: PublicKey -> PublicKey -> Bool

(<=) :: PublicKey -> PublicKey -> Bool

(>) :: PublicKey -> PublicKey -> Bool

(>=) :: PublicKey -> PublicKey -> Bool

max :: PublicKey -> PublicKey -> PublicKey

min :: PublicKey -> PublicKey -> PublicKey

Ord PublicKey
Since PublicKey doesn't expose many instances, we convert to Lexeme and compare the results

Instance details

Methods

compare :: PublicKey -> PublicKey -> Ordering

(<) :: PublicKey -> PublicKey -> Bool

(<=) :: PublicKey -> PublicKey -> Bool

(>) :: PublicKey -> PublicKey -> Bool

(>=) :: PublicKey -> PublicKey -> Bool

max :: PublicKey -> PublicKey -> PublicKey

min :: PublicKey -> PublicKey -> PublicKey

Ord PublicKey
Since PublicKey doesn't expose many instances, we convert to Lexeme and compare the results

Instance details

Methods

compare :: PublicKey -> PublicKey -> Ordering

(<) :: PublicKey -> PublicKey -> Bool

(<=) :: PublicKey -> PublicKey -> Bool

(>) :: PublicKey -> PublicKey -> Bool

(>=) :: PublicKey -> PublicKey -> Bool

max :: PublicKey -> PublicKey -> PublicKey

min :: PublicKey -> PublicKey -> PublicKey

Ord PublicKey
Since PublicKey doesn't expose many instances, we convert to Lexeme and compare the results

Instance details

Methods

compare :: PublicKey -> PublicKey -> Ordering

(<) :: PublicKey -> PublicKey -> Bool

(<=) :: PublicKey -> PublicKey -> Bool

(>) :: PublicKey -> PublicKey -> Bool

(>=) :: PublicKey -> PublicKey -> Bool

max :: PublicKey -> PublicKey -> PublicKey

min :: PublicKey -> PublicKey -> PublicKey

Read Signature

Instance details

Methods

readsPrec :: Int -> ReadS Signature

readList :: ReadS [Signature]

readPrec :: ReadPrec Signature

readListPrec :: ReadPrec [Signature]

Read PublicKey

Instance details

Methods

readsPrec :: Int -> ReadS PublicKey

readList :: ReadS [PublicKey]

readPrec :: ReadPrec PublicKey

readListPrec :: ReadPrec [PublicKey]

Read Address

Instance details

Methods

readsPrec :: Int -> ReadS Address

readList :: ReadS [Address]

readPrec :: ReadPrec Address

readListPrec :: ReadPrec [Address]

ToJSONKey PublicKey

Instance details

Methods

toJSONKey :: ToJSONKeyFunction PublicKey

toJSONKeyList :: ToJSONKeyFunction [PublicKey]

FromJSONKey PublicKey

Instance details

Methods

fromJSONKey :: FromJSONKeyFunction PublicKey

fromJSONKeyList :: FromJSONKeyFunction [PublicKey]

IsoCValue (CValue t)

Instance details

Associated Types

type ToCT (CValue t) :: CT

Methods

toCVal :: CValue t -> CValue (ToCT (CValue t))

fromCVal :: CValue (ToCT (CValue t)) -> CValue t

IsoValue (Value' Instr t)

Instance details

Associated Types

type ToT (Value' Instr t) :: T

Methods

toVal :: Value' Instr t -> Value (ToT (Value' Instr t))

fromVal :: Value (ToT (Value' Instr t)) -> Value' Instr t

(Read a, KnownSymbol name) => Read (NamedF Identity a name)

Instance details

Methods

readsPrec :: Int -> ReadS (NamedF Identity a name)

readList :: ReadS [NamedF Identity a name]

readPrec :: ReadPrec (NamedF Identity a name)

readListPrec :: ReadPrec [NamedF Identity a name]