Michelson.Typed.Annotation.Path

Documentation

type EpPath = Path Symbol

An entrypoint-Path is a Path where the annotations are Symbol's

data Path a where

A Path a is an extended entrypoint over Michelson types Annotated with a

Constructors

(:*) :: Path a -> Path a -> Path a

(:+) :: a -> Path a -> Path a

Here :: Path a

ppPath :: Path Text -> Text

Pretty-print a Path Text

type (:*@#@$$$) (t6989586621679503956 :: Path a6989586621679498572) (t6989586621679503957 :: Path a6989586621679498572) = (:*) t6989586621679503956 t6989586621679503957

data (:*@#@$$) (t6989586621679503956 :: Path a6989586621679498572) :: (~>) (Path a6989586621679498572) (Path a6989586621679498572) where

Constructors

(::*@#@$$###) :: forall t6989586621679503956 t6989586621679503957 arg. SameKind (Apply ((:*@#@$$) t6989586621679503956) arg) ((:*@#@$$$) t6989586621679503956 arg) => (:*@#@$$) t6989586621679503956 t6989586621679503957

data (:*@#@$) :: forall a6989586621679498572. (~>) (Path a6989586621679498572) ((~>) (Path a6989586621679498572) (Path a6989586621679498572)) where

Constructors

(::*@#@$###) :: forall t6989586621679503956 arg. SameKind (Apply (:*@#@$) arg) ((:*@#@$$) arg) => (:*@#@$) t6989586621679503956

type (:+@#@$$$) (t6989586621679503960 :: a6989586621679498573) (t6989586621679503961 :: Path a6989586621679498573) = (:+) t6989586621679503960 t6989586621679503961

data (:+@#@$$) (t6989586621679503960 :: a6989586621679498573) :: (~>) (Path a6989586621679498573) (Path a6989586621679498573) where

Constructors

(::+@#@$$###) :: forall t6989586621679503960 t6989586621679503961 arg. SameKind (Apply ((:+@#@$$) t6989586621679503960) arg) ((:+@#@$$$) t6989586621679503960 arg) => (:+@#@$$) t6989586621679503960 t6989586621679503961

data (:+@#@$) :: forall a6989586621679498573. (~>) a6989586621679498573 ((~>) (Path a6989586621679498573) (Path a6989586621679498573)) where

Constructors

(::+@#@$###) :: forall t6989586621679503960 arg. SameKind (Apply (:+@#@$) arg) ((:+@#@$$) arg) => (:+@#@$) t6989586621679503960

type HereSym0 = Here

type SPath = (Sing :: Path a -> Type)

type family Equals_6989586621679505344 (a :: Path a) (b :: Path a) :: Bool where ...

Equations

Equals_6989586621679505344 ((:*) a a) ((:*) b b) = (&&) ((==) a b) ((==) a b)

Equals_6989586621679505344 ((:+) a a) ((:+) b b) = (&&) ((==) a b) ((==) a b)

Equals_6989586621679505344 Here Here = TrueSym0

Equals_6989586621679505344 (_ :: Path a) (_ :: Path a) = FalseSym0

type family Compare_6989586621679506957 (a :: Path a) (a :: Path a) :: Ordering where ...

Equations

Compare_6989586621679506957 ((:*) a_6989586621679505707 a_6989586621679505709) ((:*) b_6989586621679505711 b_6989586621679505713) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679505707) b_6989586621679505711)) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679505709) b_6989586621679505713)) '[]))

Compare_6989586621679506957 ((:+) a_6989586621679505715 a_6989586621679505717) ((:+) b_6989586621679505719 b_6989586621679505721) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679505715) b_6989586621679505719)) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679505717) b_6989586621679505721)) '[]))

Compare_6989586621679506957 Here Here = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) '[]

Compare_6989586621679506957 ((:*) _ _) ((:+) _ _) = LTSym0

Compare_6989586621679506957 ((:*) _ _) Here = LTSym0

Compare_6989586621679506957 ((:+) _ _) ((:*) _ _) = GTSym0

Compare_6989586621679506957 ((:+) _ _) Here = LTSym0

Compare_6989586621679506957 Here ((:*) _ _) = GTSym0

Compare_6989586621679506957 Here ((:+) _ _) = GTSym0

type Compare_6989586621679506957Sym2 (a6989586621679506955 :: Path a6989586621679498571) (a6989586621679506956 :: Path a6989586621679498571) = Compare_6989586621679506957 a6989586621679506955 a6989586621679506956

data Compare_6989586621679506957Sym1 (a6989586621679506955 :: Path a6989586621679498571) :: (~>) (Path a6989586621679498571) Ordering where

Constructors

Compare_6989586621679506957Sym1KindInference :: forall a6989586621679506955 a6989586621679506956 arg. SameKind (Apply (Compare_6989586621679506957Sym1 a6989586621679506955) arg) (Compare_6989586621679506957Sym2 a6989586621679506955 arg) => Compare_6989586621679506957Sym1 a6989586621679506955 a6989586621679506956

data Compare_6989586621679506957Sym0 :: forall a6989586621679498571. (~>) (Path a6989586621679498571) ((~>) (Path a6989586621679498571) Ordering) where

Constructors

Compare_6989586621679506957Sym0KindInference :: forall a6989586621679506955 arg. SameKind (Apply Compare_6989586621679506957Sym0 arg) (Compare_6989586621679506957Sym1 arg) => Compare_6989586621679506957Sym0 a6989586621679506955

type family ShowsPrec_6989586621679508964 (a :: Nat) (a :: Path a) (a :: Symbol) :: Symbol where ...

Equations

ShowsPrec_6989586621679508964 p_6989586621679508291 ((:*) argL_6989586621679508293 argR_6989586621679508295) a_6989586621679508946 = Apply (Apply (Apply ShowParenSym0 (Apply (Apply (>@#@$) p_6989586621679508291) (FromInteger 9))) (Apply (Apply (.@#@$) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argL_6989586621679508293)) (Apply (Apply (.@#@$) (Apply ShowStringSym0 (FromString " :* "))) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argR_6989586621679508295)))) a_6989586621679508946

ShowsPrec_6989586621679508964 p_6989586621679508291 ((:+) argL_6989586621679508297 argR_6989586621679508299) a_6989586621679508948 = Apply (Apply (Apply ShowParenSym0 (Apply (Apply (>@#@$) p_6989586621679508291) (FromInteger 9))) (Apply (Apply (.@#@$) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argL_6989586621679508297)) (Apply (Apply (.@#@$) (Apply ShowStringSym0 (FromString " :+ "))) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argR_6989586621679508299)))) a_6989586621679508948

ShowsPrec_6989586621679508964 _ Here a_6989586621679508950 = Apply (Apply ShowStringSym0 (FromString "Here")) a_6989586621679508950

type ShowsPrec_6989586621679508964Sym3 (a6989586621679508961 :: Nat) (a6989586621679508962 :: Path a6989586621679498571) (a6989586621679508963 :: Symbol) = ShowsPrec_6989586621679508964 a6989586621679508961 a6989586621679508962 a6989586621679508963

data ShowsPrec_6989586621679508964Sym2 (a6989586621679508961 :: Nat) (a6989586621679508962 :: Path a6989586621679498571) :: (~>) Symbol Symbol where

Constructors

ShowsPrec_6989586621679508964Sym2KindInference :: forall a6989586621679508961 a6989586621679508962 a6989586621679508963 arg. SameKind (Apply (ShowsPrec_6989586621679508964Sym2 a6989586621679508961 a6989586621679508962) arg) (ShowsPrec_6989586621679508964Sym3 a6989586621679508961 a6989586621679508962 arg) => ShowsPrec_6989586621679508964Sym2 a6989586621679508961 a6989586621679508962 a6989586621679508963

data ShowsPrec_6989586621679508964Sym1 (a6989586621679508961 :: Nat) :: forall a6989586621679498571. (~>) (Path a6989586621679498571) ((~>) Symbol Symbol) where

Constructors

ShowsPrec_6989586621679508964Sym1KindInference :: forall a6989586621679508961 a6989586621679508962 arg. SameKind (Apply (ShowsPrec_6989586621679508964Sym1 a6989586621679508961) arg) (ShowsPrec_6989586621679508964Sym2 a6989586621679508961 arg) => ShowsPrec_6989586621679508964Sym1 a6989586621679508961 a6989586621679508962

data ShowsPrec_6989586621679508964Sym0 :: forall a6989586621679498571. (~>) Nat ((~>) (Path a6989586621679498571) ((~>) Symbol Symbol)) where

Constructors

ShowsPrec_6989586621679508964Sym0KindInference :: forall a6989586621679508961 arg. SameKind (Apply ShowsPrec_6989586621679508964Sym0 arg) (ShowsPrec_6989586621679508964Sym1 arg) => ShowsPrec_6989586621679508964Sym0 a6989586621679508961