Michelson.Typed.EntryPoints.Sing.Alg

Documentation

newtype EpValueF (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) where

EpValue where the fields may be empty.

Constructors

EpValueF :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t). NP (EpFields f t ann) (EpPaths ann) -> EpValueF f t ann

emptyEpValueF :: forall f (t :: TAlg) (ann :: SymAnn t). AltError [String] f => Sing t -> Sing ann -> EpValueF f t ann

An empty EpValueF, using emptyEpFields

transEpValueF :: forall (f :: Type -> Type) (g :: Type -> Type) (t :: TAlg) (ann :: SymAnn t). (forall t'. SingI t' => f (ValueOpq t') -> g (ValueOpq t')) -> Sing t -> Sing ann -> EpValueF f t ann -> EpValueF g t ann

Transform the f in EpValueF

lensEpValueF :: forall f (t :: TAlg) (ann :: SymAnn t). (Show1 f, AltError [String] f) => Sing t -> Sing ann -> Lens' (ValueAlgT f t) (EpValueF f t ann)

prfAllShow :: forall a g (xs :: [a]). (HasDict1 a, forall x. SingI x => Show (g x)) => Sing xs -> Dict (All (Compose Show g) xs)

A proof that SingI implies Show implies All Show

npToNS :: forall a (f :: Type -> Type) (g :: a -> Type) (xs :: [a]). (HasDict1 a, AltError [String] f, forall x. SingI x => Show (g x)) => Sing xs -> NP (f :.: g) xs -> f (NS g xs)

Convert NP to NS by picking the only non-erroring entrypoint as the sum choice

emptyNP :: forall a (f :: Type -> Type) (g :: a -> Type) (xs :: [a]). (HasDict1 a, SingKind a, Show (Demote a), AltError [String] f, forall x. SingI x => Show (g x), SingI xs, SListI xs) => NP (f :.: g) xs

Use altErr to make an empty NP

nsToNP :: forall a (f :: Type -> Type) (g :: a -> Type) (xs :: [a]). (HasDict1 a, SingKind a, Show (Demote a), AltError [String] f, forall x. Show x => Show (f x), forall x. SingI x => Show (g x), SingI xs, SListI xs) => NS g xs -> NP (f :.: g) xs

Convert NS to NP by filling all non-chosen slots with altErr and running pure on the single chosen slot

newtype EpValue (t :: TAlg) (ann :: SymAnn t) where

Constructors

EpValue :: forall (t :: TAlg) (ann :: SymAnn t). NS (EpFields I t ann) (EpPaths ann) -> EpValue t ann

runEpValue :: forall f (t :: TAlg) (ann :: SymAnn t). (AltError [String] f, Monad f, Show1 f, forall x. Show x => Show (f x)) => Sing t -> Sing ann -> EpValue t ann -> f (Value (FromTAlg t))

Convert an EpValue to the Value it represents

fromEpValueF :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t). AltError [String] f => Sing t -> Sing ann -> EpValueF f t ann -> f (EpValue t ann)

Assert that no fields error and convert to an EpValue

prfAllShowEpFields :: forall t (ann :: SymAnn t) xs. Sing t -> Sing ann -> Sing xs -> Dict (All (Compose Show (EpFields I t ann)) xs)

toEpValueF :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t). (AltError [String] f, Show1 f, forall x. Show x => Show (f x)) => Sing t -> Sing ann -> EpValue t ann -> EpValueF f t ann

Convert an EpValue to an EpValueF

epValueFields :: forall (t :: TAlg) (ann :: SymAnn t) r. (forall (epPath :: EpPath). Sing epPath -> RunAltE WrappedSing (NP (EpField I t ann epPath)) (EpFieldNamesErrM t ann epPath) -> r) -> Sing t -> Sing ann -> EpValue t ann -> r

Extract the fields from an EpValue