Michelson.Typed.EntryPoints.Sing.Alg.Field

Documentation

singAllSingI :: forall a (xs :: [a]). HasDict1 a => Sing xs -> Dict (All SingI xs)

Sing implies All SingI

npWrappedSing :: forall a (xs :: [a]). HasDict1 a => Sing xs -> NP WrappedSing xs

Produce a NP from a Sing list by exposing the Sing elements with WrappedSing

data EpField (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol) where

A single field of a value

Constructors

EpField :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol). Sing fieldName -> RunSingValueOpq f (EpFieldT t ann epPath fieldName) -> EpField f t ann epPath fieldName

unwrapEpField :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol). Applicative f => EpField f t ann epPath fieldName -> f (EpField I t ann epPath fieldName)

Expose the effects of f at the top level of an EpField

wrapEpField :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol). (Functor f, SingI epPath, SingI fieldName) => Sing t -> Sing ann -> f (EpField I t ann epPath fieldName) -> EpField f t ann epPath fieldName

Embed the effects of f at the top level of an EpField

emptyEpField :: forall (f :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol). AltError [String] f => Sing t -> Sing ann -> Sing epPath -> Sing fieldName -> EpField f t ann epPath fieldName

Make an empty EpField using altErr

transEpField :: forall (f :: Type -> Type) (g :: Type -> Type) (t :: TAlg) (ann :: SymAnn t) (epPath :: EpPath) (fieldName :: Maybe Symbol). (forall t'. SingI t' => f (ValueOpq t') -> g (ValueOpq t')) -> Sing t -> Sing ann -> Sing epPath -> EpField f t ann epPath fieldName -> EpField g t ann epPath fieldName

Transform the f of an EpField