Michelson.Typed.Annotation.Path

Documentation

type EpPath = Path Symbol

An entrypoint-Path is a Path where the annotations are Symbol's

data Path a where

A Path a is an extended entrypoint over Michelson types Annotated with a

Constructors

(:*) :: Path a -> Path a -> Path a

(:+) :: a -> Path a -> Path a

Here :: Path a

ppPath :: Path Text -> Text

Pretty-print a Path Text

type (:*@#@$$$) (t6989586621679507246 :: Path a6989586621679501902) (t6989586621679507247 :: Path a6989586621679501902) = (:*) t6989586621679507246 t6989586621679507247

data (:*@#@$$) (t6989586621679507246 :: Path a6989586621679501902) :: (~>) (Path a6989586621679501902) (Path a6989586621679501902) where

Constructors

(::*@#@$$###) :: forall t6989586621679507246 t6989586621679507247 arg. SameKind (Apply ((:*@#@$$) t6989586621679507246) arg) ((:*@#@$$$) t6989586621679507246 arg) => (:*@#@$$) t6989586621679507246 t6989586621679507247

data (:*@#@$) :: forall a6989586621679501902. (~>) (Path a6989586621679501902) ((~>) (Path a6989586621679501902) (Path a6989586621679501902)) where

Constructors

(::*@#@$###) :: forall t6989586621679507246 arg. SameKind (Apply (:*@#@$) arg) ((:*@#@$$) arg) => (:*@#@$) t6989586621679507246

type (:+@#@$$$) (t6989586621679507250 :: a6989586621679501903) (t6989586621679507251 :: Path a6989586621679501903) = (:+) t6989586621679507250 t6989586621679507251

data (:+@#@$$) (t6989586621679507250 :: a6989586621679501903) :: (~>) (Path a6989586621679501903) (Path a6989586621679501903) where

Constructors

(::+@#@$$###) :: forall t6989586621679507250 t6989586621679507251 arg. SameKind (Apply ((:+@#@$$) t6989586621679507250) arg) ((:+@#@$$$) t6989586621679507250 arg) => (:+@#@$$) t6989586621679507250 t6989586621679507251

data (:+@#@$) :: forall a6989586621679501903. (~>) a6989586621679501903 ((~>) (Path a6989586621679501903) (Path a6989586621679501903)) where

Constructors

(::+@#@$###) :: forall t6989586621679507250 arg. SameKind (Apply (:+@#@$) arg) ((:+@#@$$) arg) => (:+@#@$) t6989586621679507250

type HereSym0 = Here

type SPath = (Sing :: Path a -> Type)

type family Equals_6989586621679508634 (a :: Path a) (b :: Path a) :: Bool where ...

Equations

Equals_6989586621679508634 ((:*) a a) ((:*) b b) = (&&) ((==) a b) ((==) a b)

Equals_6989586621679508634 ((:+) a a) ((:+) b b) = (&&) ((==) a b) ((==) a b)

Equals_6989586621679508634 Here Here = TrueSym0

Equals_6989586621679508634 (_ :: Path a) (_ :: Path a) = FalseSym0

type family Compare_6989586621679510239 (a :: Path a) (a :: Path a) :: Ordering where ...

Equations

Compare_6989586621679510239 ((:*) a_6989586621679508997 a_6989586621679508999) ((:*) b_6989586621679509001 b_6989586621679509003) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679508997) b_6989586621679509001)) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679508999) b_6989586621679509003)) '[]))

Compare_6989586621679510239 ((:+) a_6989586621679509005 a_6989586621679509007) ((:+) b_6989586621679509009 b_6989586621679509011) = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679509005) b_6989586621679509009)) (Apply (Apply (:@#@$) (Apply (Apply CompareSym0 a_6989586621679509007) b_6989586621679509011)) '[]))

Compare_6989586621679510239 Here Here = Apply (Apply (Apply FoldlSym0 ThenCmpSym0) EQSym0) '[]

Compare_6989586621679510239 ((:*) _ _) ((:+) _ _) = LTSym0

Compare_6989586621679510239 ((:*) _ _) Here = LTSym0

Compare_6989586621679510239 ((:+) _ _) ((:*) _ _) = GTSym0

Compare_6989586621679510239 ((:+) _ _) Here = LTSym0

Compare_6989586621679510239 Here ((:*) _ _) = GTSym0

Compare_6989586621679510239 Here ((:+) _ _) = GTSym0

type Compare_6989586621679510239Sym2 (a6989586621679510237 :: Path a6989586621679501901) (a6989586621679510238 :: Path a6989586621679501901) = Compare_6989586621679510239 a6989586621679510237 a6989586621679510238

data Compare_6989586621679510239Sym1 (a6989586621679510237 :: Path a6989586621679501901) :: (~>) (Path a6989586621679501901) Ordering where

Constructors

Compare_6989586621679510239Sym1KindInference :: forall a6989586621679510237 a6989586621679510238 arg. SameKind (Apply (Compare_6989586621679510239Sym1 a6989586621679510237) arg) (Compare_6989586621679510239Sym2 a6989586621679510237 arg) => Compare_6989586621679510239Sym1 a6989586621679510237 a6989586621679510238

data Compare_6989586621679510239Sym0 :: forall a6989586621679501901. (~>) (Path a6989586621679501901) ((~>) (Path a6989586621679501901) Ordering) where

Constructors

Compare_6989586621679510239Sym0KindInference :: forall a6989586621679510237 arg. SameKind (Apply Compare_6989586621679510239Sym0 arg) (Compare_6989586621679510239Sym1 arg) => Compare_6989586621679510239Sym0 a6989586621679510237

type family ShowsPrec_6989586621679511617 (a :: Nat) (a :: Path a) (a :: Symbol) :: Symbol where ...

Equations

ShowsPrec_6989586621679511617 p_6989586621679511565 ((:*) argL_6989586621679511567 argR_6989586621679511569) a_6989586621679511599 = Apply (Apply (Apply ShowParenSym0 (Apply (Apply (>@#@$) p_6989586621679511565) (FromInteger 9))) (Apply (Apply (.@#@$) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argL_6989586621679511567)) (Apply (Apply (.@#@$) (Apply ShowStringSym0 (FromString " :* "))) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argR_6989586621679511569)))) a_6989586621679511599

ShowsPrec_6989586621679511617 p_6989586621679511565 ((:+) argL_6989586621679511571 argR_6989586621679511573) a_6989586621679511601 = Apply (Apply (Apply ShowParenSym0 (Apply (Apply (>@#@$) p_6989586621679511565) (FromInteger 9))) (Apply (Apply (.@#@$) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argL_6989586621679511571)) (Apply (Apply (.@#@$) (Apply ShowStringSym0 (FromString " :+ "))) (Apply (Apply ShowsPrecSym0 (FromInteger 10)) argR_6989586621679511573)))) a_6989586621679511601

ShowsPrec_6989586621679511617 _ Here a_6989586621679511603 = Apply (Apply ShowStringSym0 (FromString "Here")) a_6989586621679511603

type ShowsPrec_6989586621679511617Sym3 (a6989586621679511614 :: Nat) (a6989586621679511615 :: Path a6989586621679501901) (a6989586621679511616 :: Symbol) = ShowsPrec_6989586621679511617 a6989586621679511614 a6989586621679511615 a6989586621679511616

data ShowsPrec_6989586621679511617Sym2 (a6989586621679511614 :: Nat) (a6989586621679511615 :: Path a6989586621679501901) :: (~>) Symbol Symbol where

Constructors

ShowsPrec_6989586621679511617Sym2KindInference :: forall a6989586621679511614 a6989586621679511615 a6989586621679511616 arg. SameKind (Apply (ShowsPrec_6989586621679511617Sym2 a6989586621679511614 a6989586621679511615) arg) (ShowsPrec_6989586621679511617Sym3 a6989586621679511614 a6989586621679511615 arg) => ShowsPrec_6989586621679511617Sym2 a6989586621679511614 a6989586621679511615 a6989586621679511616

data ShowsPrec_6989586621679511617Sym1 (a6989586621679511614 :: Nat) :: forall a6989586621679501901. (~>) (Path a6989586621679501901) ((~>) Symbol Symbol) where

Constructors

ShowsPrec_6989586621679511617Sym1KindInference :: forall a6989586621679511614 a6989586621679511615 arg. SameKind (Apply (ShowsPrec_6989586621679511617Sym1 a6989586621679511614) arg) (ShowsPrec_6989586621679511617Sym2 a6989586621679511614 arg) => ShowsPrec_6989586621679511617Sym1 a6989586621679511614 a6989586621679511615

data ShowsPrec_6989586621679511617Sym0 :: forall a6989586621679501901. (~>) Nat ((~>) (Path a6989586621679501901) ((~>) Symbol Symbol)) where

Constructors

ShowsPrec_6989586621679511617Sym0KindInference :: forall a6989586621679511614 arg. SameKind (Apply ShowsPrec_6989586621679511617Sym0 arg) (ShowsPrec_6989586621679511617Sym1 arg) => ShowsPrec_6989586621679511617Sym0 a6989586621679511614